Refactoring identifier logic into coins.
Ask known coin types if they match, starting with dimes.
I like this approach, and the tests still pass. Now I want
to add nickels and quarters and use a loop for all Coins.

I am wishing I had an existing "invalid" test, however. It's
awkward wanting to have a loop, which means it is possible
that none of them would return a type.